Search duplicate resources in clustermap when partition cant be found in given resource (#2626)
In helix, when we add new resources to ideal state, it will take some time to populate that information to aggregated external view. Within this time, helix cluster manager would know the resources for partitions, but getReplicaIdsByState would fail because the routing table hasn't updated yet.

This PR would fix this issue by searching in old resources to find the replica ids.