Use clustermap to avoid state transitions from old resource (#2631)
When reconstructing resources, if some hosts are down, and when they wake up, they would receive state transition messages for the old resources, which is to drop partitions. At the time, if the message of old resources comes before the message of new resources, ambry would drop those partitions. This is not good.

This PR fixes that by relying clustermap to provide the resource for a partition. If the resource provided by the clustermap is higher than the resource from the message, we would ignore this transition.