Fix NPE in helix cluster manager (#2648)
When hosts are removed from cluster, it looks like there is some race condition in helix callbacks which causes the hosts to be present under /DataNodeConfigs but their znode itself is missing. This results in exception and causes the in-memory cluster map to be incomplete. Adding null checks to prevent this.