[vcr-2.0] Fix Compaction schedule (#2643)
A problem with cloud-compaction is that it is scheduled at a fixed intervals or rate. These intervals are arbitrary and we cannot determine the time it takes for a single run of cloud-compaction to finish. If a compaction run takes longer than the fixed-interval, then the fixed-rate scheduling starts the next run immediately after the previous one finishes to keep up the rate. A better approach is to let compaction take its time and introduce a sleep between consecutive cycles so as to not waste resources.

Lastly, this patch simplifies the logic for scheduling compaction workers by relying just on Java's in-built ExecutorService, and no other logic. It uses proven Util methods and daemon threads to ensure the server exits on a shutdown.

There is still one issue with cloud-compaction. If a partition is removed and reassigned to another host, then compaction for that partition could be running on two hosts at the same time. Though not harmful, it just wastes network resources.