[HelixTask] Reduce memory footprint to minimize GC pauses (#2657)
The frontends are timing out connecting to servers because servers are busy with GC. The GC pause is 3-second long because JVM is busy releasing memory consumed by StatsAggregator. The aggregator fetches results for all accounts, containers and partitions to store it in a giant map, which consumes close to 10GB of heap.

This patch fixes this issue by not buffering those stats. Instead it fetches each stat and immediately processes it without saving it for further use.