Include leaderless replicas for GET replication requests (#2661)
During leader-based replication, we only include leaders for fetching actual data. For non-leaders, we wait for 2 minutes hoping that the data would be fetched via intra-colo replication. This causes issue of very slow replication if leader is missing. To avoid that, we are fetching data for replicas without any leaders in this PR.