Fail storage manager when too many disk failed (#2690)
Fail the initialization of StorageManager when there are too many disks fail to start or marked as unavailable. If enough disks fail to start up, there is no way ambry-server would be able to function. Then just fail the ambry-server so the helix would move all the replicas out sometime.