Fix a NPE in getReplicaIdsByState (#2695)
When new resource is created for the exiting partitions, the ideal state would have this partition before the external view. So checking if the new external view has any replica of this partition has to make sure that partition state map really exist. If it does not exist, then we just have to return false right away.