[Recovery] Impl get-metadata from azure-storage (#2687)
This patch implements the basic functionality required to initiate recovery of data from Azure. The recovery code will run directly on the server, which can contact azure and fetch blobs. The first step is to fetch metadata. Azure Storage provides a continuation token which the server can use to fetch paginated metadata. When the server runs in recovery mode, it does not replicate from its peer servers. It only replicates data from Azure. Recovery for a partition is complete when the token issued by Azure is null. This patch uses a json format to store the token along with some additional info for convenience.

The major files are RecoveryNetworkClient and RecoveryNetworkClientTest