Fix ambry server infinity stacktrace when netty channel is closed by exception (#2691)
When there is an exception in storage server netty layer, it will close the channel, which would try to flush any bytes on the channel, which would then trigger the exception again, and then calling the handler to close the channel again. Then we are trapped in an infinite stack traces.

Unfortunately, jvm doesn't kill the process when this happens, so the ambry-server process is returning healthy to health checker while refusing to serve any requests.

This PR fixes that.