[vcr-2.0] Introduce CPU scaling factor (#2741)
Not all nodes are identical; some possess more cores while others fewer.
Therefore, adopting a predetermined count of replication threads is not advisable.
Instead, the number of replication threads should be determined based on the number of cores available.
The scaling factor provided below can assume three potential states: negative, zero, or positive.
A negative value signifies the creation of a number of threads equal to the absolute value of the negative integer.
Zero denotes the absence of thread creation.
A positive value will serve as a multiplier for the number of CPUs to ascertain the number of threads to generate.
For eg, for a node with 4 cores,

A cpu_scale of -2.5 will create 2 threads since int(|-2.5|) = 2
A cpu_scale of +2.5 will create 10 threads since int(2.5 * 4) = 10