Add a new metric to show the number of full range compaction (#2759)
Adding a new metric to show how many full range compaction is going on each host. Full range compaction takes in all the inactive log segment files and run compaction logic on top of them. It's a very slow process and consumes a lot of disk iops. Having a metric to show the number of full range compaction will provide us with more insight to the cluster.