Filter out all valid log semgents in full range compaction (#2773)
In full range compaction (CompactAllPolicy), we would return all log segments what are not in journal to compactor, even if some of the log segments are having 100% of valid data. This is totally unnecessary, especially for leading and trailing log segments. Since we would have to copy all of their data out to a new log segment file.