[DR] Compare metadata from all replicas of a partition with its backup (#2787)
This patch adds support in the verification-app to compare blob metadata between all 3 server replicas of a partition and its cloud backup. It also adds some filters to reduce false alarms and unhelpful comparisons.
Lastly, we leave the partition on the local disk of the verification app to allow for debugging, instead of wiping it out. It will be wiped out later anyways.