[DR] Debug branch
This patch helps debug the backup-checker-app to rule out any bugs
in the code or debug any issues in the backup.

To use this patch, pull it, rebase on master.
Apply your changes, if any.
Compile and copy the necessary jar files on to the backup-checker hosts.

Usecases:
1. Restoring a selected partition

If you wish to recover a particular partition from cloud and run
checker on it, you can do so with this patch.

2. Re-using restored partition from local-disk
Once the backup has been restored, you'd want to use the copy in the
local-disk of the backup-checker host and re-run checker for
further troubleshooting. But each time the backup-checker-app starts,
it wipes the disk. This patch prevents that allowing you to debug.