[DR] Get blob-content CRC of Azure blobs (#2795)
We are going to compare CRC of blob from server and from Azure. This patch adds code to get blob-content CRC of Azure blobs recovered and stored in the local disk of backup-verification-app,. Refactored the code so that the same method runs in server and in verification-app to fetch CRC from local blob store.