Fix a bug when adding new partitions (#2800)
When adding a new partition to a new host in FULL AUTO, we would wait go through helix state transition to create a bootstrap replica before we update the property store to add replicas to data node config. When creating bootstrap replica, we would put the partition name to partitionNameToAmbryParitition map, so later when the property store is updated, we would ignore adding partition to partitionMap. That's why we encountered a Partition Id from stream is unknown error.

This pr fixes that.