[vcr-2.0] Tune thread-local-cache (#2804)
The memory load is quite high on backup-nodes and the vcr-server is crashing due to out-of-memory error. I do not suspect a memleak. The crash is happening after enabling the thread-local-cache, which is just a wrapper over open-src Caffeine cache. The cache is capped by the number of objects, but I suspect that even though the cache itself is not full, there isn't any memory to insert another item in the cache.