[vcr-2.0] Remove unnecessary sleep in azure-client (#2830)
The Azure SDK interprets the tryTimeoutInSeconds as sleep duration after each request to Azure. However, we are using a sync-client which blocks for the same duration waiting on a response. So we don't need an additional sleep in the client due to tryTimeoutInSeconds.

The above is azure-blob-client. Also, fixes azure-table-client similarly.