[vcr-2.0] Remove unused and excessive metrics (#2845)
VCR creates a lot of metrics per partition and none of these are used. Just get rid of it. We have way too many metrics causing errors when sending that message to Kafka. We have an open ticket with Kafka regarding this but the recommendation is also to reduce metrics on our side.