[vcr-2.0] Delay replica-thread start and token init (#2866)
There is a burst of calls to Azure to reload replica tokens when a node starts because helix adds replicas very quickly. Due to this, there are a lot of errors resulting in a token reset. To avoid this, let's add a small random delay before a replica thread starts and delay the token initialization until necessary.

Also remove a bunch of unwanted configs and code.