Use client error when client closed connection (#2868)
When client closes the connection while downloading a blob, we will set the error response status to InternalServerError and then increment a GetBlobServerError counter in ContainerMetrics. But this is not a server error, this is a client.

This PR fixes this discrepancy. We will set error response status to client error so later it won't be counted as server error.