[vcr-2.0] Split blob-compaction metrics from blob-delete metrics (#2865)
The blob-deletion code path implementing the IMMEDIATE blob-deletion policy reuses the existing code for compaction to erase a blob. This code was originally designed for compaction and includes compaction-specific metrics. When IMMEDIATE blob-deletion is enabled, it appears as though no blob deletions are occurring because the metric num-blob-delete is only incremented when EVENTUAL blob-deletion is active. This patch addresses the confusion by fixing the metric handling and correcting some typos in the metric names.

This patch also removes grace-period for holding expired blobs, to match what the server does.