[vcr-2.0] Fix duplicate uploads (#2869)
We were attempting duplicate uploads because we were assigning replicas of a partition to different threads, which resulted in multiple threads trying to upload the same blob simultaneously. This behavior was due to a legacy code artifact. This patch corrects these issues and includes guard code to prevent future occurrences and facilitate debugging if it happens again.