[Paranoid Durability] Limit the number of remote PUT attempts. (#2876)
This PR adds a new config and code for limiting the number of times we attempt to PUT to a remote colo. The intention is to limit the memory usage during times of extra high network latency. When the network is slow or failing, ambry-frontend will hold request attempts in memory, which can lead to exhaustion of Java direct memory.

If we reach the limit for remote writes, but local writes are successful, we will still return success to the client. The assumption here is that it is better to degrade performance than to completely fail, since we assume that network latency/error spikes will be fairly brief.

A new metric and log message is added to allow debugging. Also, an old log message is removed because it was found to make the logs noisy without adding much value.