[vcr-2.0] Measure Azure Storage Container lag in bytes (#2884)
This patch measures the drift of a Azure Storage container from a Ambry partition in bytes. It introduces a class responsible for aggregating container metrics. We deliberately avoid emitting per-container metrics to prevent an overwhelming increase in the number of metrics, which would strain the telemetry system. This approach has been tried previously and was unsuccessful.

A daemon will run periodically, emitting aggregate metrics in a controlled and predictable manner. The drift of the Azure container from the Ambry partition is set using a compare-and-set mechanism to prevent accidental multithreading errors. However, it's unlikely that multiple threads will handle the same partition in the VCR, as a single thread manages all replicas of a partition using ROUND_ROBIN policy.

We use the min() function because bootstrapping replicas can skew the data, falsely indicating a large drift when the partition is fully backed up. If the lag or drift is -1, we round it up to 0 and proceed.