Remove initial delay in loading json-accounts file (#3026)
This unpredictable initial delay is frustrating because we can't issue any commands to Ambry until the accounts file is loaded, and we have no way of knowing when that will happen. Additionally, this delay is linked to the regular polling delay configuration, making it hard to adjust. Since I don’t want to add a separate configuration for the initial delay, I’m simply setting it to 0.