Recover from last log segment with invalid messages (#3018)
Since we don't ignore invalid log segment from this PR #2744, we run into lots of error replicas due to invalid messages in last log segment when power outage happens. This PR adds configuration and logic to deal with those errors.

When there are invalid messages in the log segment when persistent index is trying to recover index segment files, we check if the log segment is the last log segment and if the amount of remaining bytes are small, if so, we just ignore the invalid message and move on with the valid message.