Add stale error to recoverable exception (#3047)
When a blob store didn't start in the last 7(by default) days, we would consider this blob store "stale" and if the StoreConfig.storeBlockStaleBlobStoreToStart is true, we would block this blobstore to start by throwing a StoreException with StoreStaleError.

This PR would try this error as recoveriable error and remove the directory for the blob store and restart the blob store automatically.