Implement InMemoryAccountService (#3072)
* Implement InMemoryAccountService

This implementation of the AccountService enables the /accounts
API to work when deploying Ambry locally without any external dependencies.
Note that this implementation is primarily meant for local testing
and not for production use.

The InMemoryAccountService essentially combines the functionality of the previous
InMemoryUnknownAccountService and JsonAccountService. However, since
the /accounts API works, it does not depend on continuously reading a
JSON file on disk to fetch account info (but does accept one on startup
to quickly fill in data).

Delete the InMemoryUnknownAccountService and JsonAccountService
since they are no longer needed with this implementation.

* review comments

* reformat

---------

Co-authored-by: Danny Cheng <dacheng@dacheng-mn2116.linkedin.biz>