Remove all versions in mysql database for delete operation (#3061)
Fix an issue with blob name delete. When deleting a named blob, we only delete the max version whose blob state is ready in the database table. And leave the other version as "undeleted". There is an issue with this approach. The background job that removes the rows whose "deleted_ts" is not null, would end up completely delete the max version row. After that, the lower versions would become the new max version as if this blob name is not deleted yet.

This PR fixed it by removing all versions.



* Fix tests

* Format