Raise clearer error when trying to POST at /named (#3079)
* Raise clearer error when trying to POST at /named

This method isn't actually supported.
When following the old README instructions for creating
a named blob, it did not specify that the method needed
to be PUT, so when I did POST I got a confusing 500
error.

* Don't break s3 post API, revert exception header change since we can
add a property in the exception to surface more info to the client

* Move failure logic to preprocessandrouterequest

* reformat

* Move check to callback

* reformat

---------

Co-authored-by: Danny Cheng <dacheng@dacheng-mn2116.linkedin.biz>