Fix potential memory leak in GetBlobOperation
Since chunkAsyncWriteCallback is passed to asyncWritableChannel we don't have guarantees
on if it is run across a single or multiple threads. In the concurrent case, it had a
get-then-increment pattern which can cause chunkIndexToBufWaitingForRelease.remove(currentNumChunk)
to skip a a chunk and cause a memory leak.

The added test will fail if the change to chunkAsyncWriteCallback is removed.