Update to Mocktio 5
Ambry currently uses Mockito 2 and Powermock (which is no longer supported).

The main reason for the upgrade is not just for maintanance but also because
Mockito 5 supports ByteBuddy 14.x which includes some fixes and functionality
improvements over the version in Mockito 2. This recent version of ByteBuddy
helps with some research being done on indentifying potential deadlocks and
memory leaks in testing, so prefer the more recent version of Mockito.

Besides, updating dependencies to the most recent releases is just good practice.