Go one step further and prevent use-after-free race condition
After even more extensive review, we identified that not only is there a case where the read buf
is leaked, it's also possible for the read buf to be used after free in rare cases where the
memory is released by the network and GCed before it is retained by the ChunkFiller thread.

Some of the memory leak in the previous commit is masking some of the use-after-free issues
fixed in this commit. Given how intertwined they are, it wouldn't be safe to merge these
changes separately.