Wrap expensive log trace calls with if enabled condition (#3185)
While this log strings and logging operations are not taken when the log level is not at the
trace level, if the parameter passed to the log statement is expensive to fetch that price is
still paid every time (depending on JIT). To make our expansive logging more performance we wrap
trace logs with a conditional check if the parameters passed are classified as potentially expensive.